refactor: product 타입 분리
