+++ src/api/product.ts
@@ -1,20 +1,5 @@
 import apiClient from "@/api/apiClient";
-
-export type Product = {
-  id: number;
-  name: string;
-  imageURL: string;
-  price: {
-    basicPrice: number;
-    sellingPrice: number;
-    discountRate: number;
-  };
-  brandInfo: {
-    id: number;
-    name: string;
-    imageURL: string;
-  };
-};
+import type { Product } from "@/types/product";
 
 export type ProductSummary = {
   id: number;
